Update 3 after review: Implement iterative average structure (Issue#2039)
